requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
imap-tools==1.0.0
email-validator==2.1.0
werkzeug==2.3.7
openpyxl==3.1.2
reportlab==4.0.4
gunicorn==21.2.0